Changed unary operator precedence
This is quite a big change, code-wise. Previously,  unary operators, even unknown ones, where given the highest precedence. Even over known binary operators that should have had a higher precedence. What I've done is I've changed how the parser and interpreter works. Now the unary operators are in a category (or in multiple categories) within the categories.catogories field. A "unary" tag is added to all unary categories. This denotes that the operators within it are unary operators. The parser has been changed so that the change in precedence is reflected, and now unary and binary operators and value nodes are joined correctly. The interpreter has been changed a little bit to reflect the merging of the two arities. The operators have been changed with the new precedences. The general unary operators have a precedence just lower than the Radix operator. All list-focused unary operators have a lower precedence than the Application and Construct operators.