Fixed bug in the standard interpreter
If an operator was in a unary category, even if the operator has a left operand, it would think that it should interpret it as a unary operator. For example, the binary operator '-' would be interpreted as the unary version. This would cause it to ignore the left operand. I fixed this by checking whether or not the root has a left operand before interpreting an operator as a unary one.