Changes packaging system
Packages are imported directly via the import_packages function. The function now returns a dict with an entry for each entry in the input dict. Each value is the requested module. The library is also given ops as an argument in shell3.py because otherwise std.ops.py was being imported twice, and the copy had seperate but identical classes, which was confusing the issubclass checks in std.int.py. This is because the type of the result of a left node child wasn't technically a subclass of the type requirement of an operator, though they were the same class.