Updated the packaging system
I have implemented memoisation in the exp_package module. This prevent double importing. Also, packages are to be imported one by one, instead of all together like they were before. Also path() has been renamed to here() because it gets the path of all things here, i.e. where exp_package is, which for it, is "here". Also I made some changes to the order which things are executed.