Added literals package
This package will be used to evaluate leaf nodes in the AST.  The LITERALS class of the package needs to have a field called literals, which is a list of pairs of a predicate and a function. The predicate will indicate if a node should be evaluated as a literal with the function.

The node will be checked against the predicates in the order they come in the list, the first one to return True will have the corresponding function applied to the node. If no predicates return true, then an UndefinedVariableError is reported.

The standard literals are in std.lit.py. It has three predicates, one for variables, one for natural numbers and one for strings.

The literals are passed to the interpreter when it is created, so that any custom alterations may be made without having to create a new literals package.