Updated variable paradigm
Also, values now have a higher priority. If a variable exists in a VARIABLE_LIST while a value also has the same name, the value is what is returned, unless get_var() is used. Also using the __getitem__ will automatically call the VARIABLE.func() method of the value or variable. Also implemented memo to prevent something from being called twice when it doesn't need to be.