Uploaded all LC stuff before hiatus.
This is all the code for my implementation for Lambda Calculus in Expript3. I'm just uploading all of this for completion's sake before I go on hiatus. Also, all the standard LC stuff has been moved to lc.ext.lib.py. This is where the LC stuff that can interact with the std operators can be found. The lc.lib will now contain a library that is only supposed to be used with lc.lit and lc.ops. This is where I have stored the literals and operators for my LC implementation.